Core TextBox implementation abstracted out to TextInputBase.  TextBox now has internal horizontal scroll to better ensure that the cursor remains in view.
